completed reason for visit module
