--- app/Http/Controllers/Admin/HomeController.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Admin;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-
-class HomeController extends Controller
-{
-    public function index()
-    {
-        return view('home');
-    }
-}
