--- app/Http/Controllers/Admin/UserController.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Admin;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-
-class UserController extends Controller
-{
-    //
-}
