--- app/Http/Controllers/VisitorController.php
@@ -1,64 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\Visitor;
-use Illuminate\Http\Request;
-
-class VisitorController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Visitor  $visitor
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Visitor $visitor)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Visitor  $visitor
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Visitor $visitor)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Visitor  $visitor
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Visitor $visitor)
-    {
-        //
-    }
-}
