+++ package.json
@@ -27,7 +27,7 @@
         "@fortawesome/fontawesome-free": "^5.15.4",
         "@johmun/vue-tags-input": "^2.1.0",
         "admin-lte": "^3.1.0",
-        "bootstrap": "^5.1.3",
+        "bootstrap": "^4.6.0",
         "laravel-vue-pagination": "^2.3.1",
         "moment": "^2.29.1",
         "sweetalert2": "^9.17.2",
