+++ resources/js/Gate.js
@@ -11,6 +11,14 @@ export default class Gate{
     isUser(){
         return this.user.type === 'user';
     }
+
+    isDoctor(){
+      return this.user.type == 'doctor';
+    }
+
+    isNurse(){
+      return this.user.type == 'nurse';
+    }
     
     isAdminOrUser(){
         if(this.user.type === 'user' || this.user.type === 'admin'){
