+++ resources/sass/_variables.scss
@@ -17,6 +17,7 @@ $yellow: #ffed4a;
 $green: #38c172;
 $teal: #4dc0b5;
 $cyan: #6cb2eb;
+$black: #000000;
 
 
 // Font Awesome
