+++ routes/api.php
@@ -39,6 +39,7 @@ Route::namespace('App\\Http\\Controllers\\API\V1')->group(function () {
         'product' => 'ProductController',
         'category' => 'CategoryController',
         'tag' => 'TagController',
-        'clients' => 'ClientController'
+        'clients' => 'ClientController',
+        'visits' => 'VisitorController'
     ]);
 });
